Add 1.2 files
Added the files for 1.2. What's new in 1.2:

Changed syntax, adding more places for comments to be added.
Re-arranged code for easier library implementation
moved scripts to make finding, editing, and adding libraries easier
added much needed commands forgotten in previous updates (And, or, xor, and the "not" versions of them all. Found in the Compare command.)